refactor(groupBy): move support types under group.ts file
