+++ src/operator/groupBy.ts
@@ -4,7 +4,6 @@ import {Observable} from '../Observable';
 import {Subject} from '../Subject';
 import {Map} from '../util/Map';
 import {FastMap} from '../util/FastMap';
-import {RefCountSubscription, GroupedObservable} from './groupBy-support';
 import {tryCatch} from '../util/tryCatch';
 import {errorObject} from '../util/errorObject';
 
@@ -136,5 +135,64 @@ class GroupDurationSubscriber<K, T> extends Subscriber<T> {
   protected _complete(): void {
     this.group.complete();
     this.parent.removeGroup(this.key);
+  }
+}
+
+export class RefCountSubscription extends Subscription {
+  primary: Subscription;
+  attemptedToUnsubscribePrimary: boolean = false;
+  count: number = 0;
+
+  constructor() {
+    super();
+  }
+
+  setPrimary(subscription: Subscription) {
+    this.primary = subscription;
+  }
+
+  unsubscribe() {
+    if (!this.isUnsubscribed && !this.attemptedToUnsubscribePrimary) {
+      this.attemptedToUnsubscribePrimary = true;
+      if (this.count === 0) {
+        super.unsubscribe();
+        this.primary.unsubscribe();
+      }
+    }
+  }
+}
+
+export class GroupedObservable<K, T> extends Observable<T> {
+  constructor(public key: K,
+              private groupSubject: Subject<T>,
+              private refCountSubscription?: RefCountSubscription) {
+    super();
+  }
+
+  protected _subscribe(subscriber: Subscriber<T>) {
+    const subscription = new Subscription();
+    if (this.refCountSubscription && !this.refCountSubscription.isUnsubscribed) {
+      subscription.add(new InnerRefCountSubscription(this.refCountSubscription));
+    }
+    subscription.add(this.groupSubject.subscribe(subscriber));
+    return subscription;
+  }
+}
+
+export class InnerRefCountSubscription extends Subscription {
+  constructor(private parent: RefCountSubscription) {
+    super();
+    parent.count++;
+  }
+
+  unsubscribe() {
+    if (!this.parent.isUnsubscribed && !this.isUnsubscribed) {
+      super.unsubscribe();
+      this.parent.count--;
+      if (this.parent.count === 0 && this.parent.attemptedToUnsubscribePrimary) {
+        this.parent.unsubscribe();
+        this.parent.primary.unsubscribe();
+      }
+    }
   }
 }
